requirements.txt
certifi==2018.4.16
Django==2.0.5
djangorestframework==3.8.2
psycopg2==2.7.4
pytz==2018.4
faker==0.8.15
